Atualização para subir heroku

+++ static/crudProdutos.js
@@ -7,7 +7,7 @@ function editProduto(produtoId, button) {
         nome: cells[0].textContent.trim(),
         preco: !isNaN(parseFloat(cells[1].textContent)) ? parseFloat(cells[1].textContent) : 0,
         quantidade: !isNaN(parseInt(cells[2].textContent)) ? parseInt(cells[2].textContent) : 0,
-        descricao: cells[3].textContent.trim(),
+        marca: cells[3].textContent.trim(),
         categoria: cells[4].textContent.trim()
     };
 
@@ -15,7 +15,7 @@ function editProduto(produtoId, button) {
     cells[0].innerHTML = `<textarea id="edit-nome" required>${originalValues.nome}</textarea>`;
     cells[1].innerHTML = `<textarea id="edit-preco" required>${originalValues.preco.toFixed(2)}</textarea>`;
     cells[2].innerHTML = `<textarea id="edit-quantidade" required>${originalValues.quantidade}</textarea>`;
-    cells[3].innerHTML = `<textarea id="edit-descricao" required>${originalValues.descricao}</textarea>`;
+    cells[3].innerHTML = `<textarea id="edit-marca" required>${originalValues.marca}</textarea>`;
     cells[4].innerHTML = `<textarea id="edit-categoria" required>${originalValues.categoria}</textarea>`;
 
     // Troca o botão "Editar" por "Salvar"
@@ -33,8 +33,9 @@ function editProduto(produtoId, button) {
             nome: document.getElementById('edit-nome').value.trim(),
             preco: parseFloat(document.getElementById('edit-preco').value),
             quantidade: parseInt(document.getElementById('edit-quantidade').value),
-            descricao: document.getElementById('edit-descricao').value.trim(),
-            categoria: document.getElementById('edit-categoria').value.trim()
+            marca: document.getElementById('edit-marca').value.trim(),
+            categoria: document.getElementById('edit-categoria').value.trim(),
+            nome_lowercase: document.getElementById('edit-nome').value.trim().toLowerCase().replace(/\s+/g, '')
         };
 
         // Validação dos campos
@@ -44,7 +45,7 @@ function editProduto(produtoId, button) {
             isNaN(updatedProduto.preco) || 
             updatedProduto.quantidade < 0 || 
             isNaN(updatedProduto.quantidade) || 
-            !updatedProduto.descricao || 
+            !updatedProduto.marca || 
             !updatedProduto.categoria
         ) {
             alert("Todos os campos devem ser preenchidos corretamente. Preço e quantidade não podem ser negativos ou nulos.");
@@ -57,30 +58,41 @@ function editProduto(produtoId, button) {
         if (changesMade) {
             const db = firebase.firestore();
             const userEmail = firebase.auth().currentUser.email;
-            const produtoRef = db.collection('users').doc(userEmail).collection('produtos').doc(produtoId);
-
-            // Atualiza o produto no Firestore
-            produtoRef.update(updatedProduto).then(() => {
-                // Atualiza a tabela com os novos valores
-                cells[0].textContent = updatedProduto.nome;
-                cells[1].textContent = updatedProduto.preco.toFixed(2);
-                cells[2].textContent = updatedProduto.quantidade;
-                cells[3].textContent = updatedProduto.descricao;
-                cells[4].textContent = updatedProduto.categoria;
-
-                // Volta o botão para "Editar"
-                saveButton.textContent = "Editar";
-                saveButton.classList.remove('save-btn');
-                saveButton.classList.add('edit-btn');
-
-                // Remove o evento de "Salvar" e volta para "Editar"
-                saveButton.removeEventListener('click', handleSaveClick);
-                saveButton.addEventListener('click', function() {
-                    editProduto(produtoId, saveButton);
-                });
-
+        
+            const distribuidoresRef = db.collection('distribuidores');
+            distribuidoresRef.where('email', '==', userEmail).limit(1).get().then((querySnapshot) => {
+                if (!querySnapshot.empty) {
+                    const distribuidorDoc = querySnapshot.docs[0];
+                    const produtoRef = distribuidorDoc.ref.collection('produtos').doc(produtoId);
+        
+                    // Atualiza o produto no Firestore
+                    produtoRef.update(updatedProduto).then(() => {
+                        // Atualiza a tabela com os novos valores
+                        cells[0].textContent = updatedProduto.nome;
+                        cells[1].textContent = updatedProduto.preco.toFixed(2);
+                        cells[2].textContent = updatedProduto.quantidade;
+                        cells[3].textContent = updatedProduto.marca;
+                        cells[4].textContent = updatedProduto.categoria;
+        
+                        // Volta o botão para "Editar"
+                        saveButton.textContent = "Editar";
+                        saveButton.classList.remove('save-btn');
+                        saveButton.classList.add('edit-btn');
+        
+                        // Remove o evento de "Salvar" e volta para "Editar"
+                        saveButton.removeEventListener('click', handleSaveClick);
+                        saveButton.addEventListener('click', function() {
+                            editProduto(produtoId, saveButton);
+                        });
+        
+                    }).catch((error) => {
+                        console.log('Erro ao atualizar produto:', error);
+                    });
+                } else {
+                    console.log('Nenhum distribuidor encontrado com o e-mail fornecido.');
+                }
             }).catch((error) => {
-                console.log('Erro ao atualizar produto:', error);
+                console.error('Erro ao buscar distribuidor:', error);
             });
         } else {
             alert("Nenhuma alteração foi realizada.");
@@ -91,18 +103,32 @@ function editProduto(produtoId, button) {
 function deleteProduto(produtoId, button) {
     const db = firebase.firestore();
     const userEmail = firebase.auth().currentUser.email;
-    const produtoRef = db.collection('users').doc(userEmail).collection('produtos').doc(produtoId);
 
-    if (confirm("Tem certeza que deseja apagar este produto?")) {
-        // Remove o produto do Firestore
-        produtoRef.delete().then(() => {
-            // Remove a linha correspondente na tabela
-            const row = button.closest('tr');
-            row.remove();
-            console.log('Produto removido com sucesso.');
-        }).catch((error) => {
-            console.log('Erro ao remover o produto:', error);
-        });
-    }
+    // Consulta para encontrar o distribuidor com o campo email igual ao userEmail
+    const distribuidoresRef = db.collection('distribuidores');
+    distribuidoresRef.where('email', '==', userEmail).limit(1).get().then((querySnapshot) => {
+        if (!querySnapshot.empty) {
+            // Pega o primeiro documento do distribuidor encontrado
+            const distribuidorDoc = querySnapshot.docs[0];
+            const produtoRef = distribuidorDoc.ref.collection('produtos').doc(produtoId);
+
+            if (confirm("Tem certeza que deseja apagar este produto?")) {
+                // Remove o produto do Firestore
+                produtoRef.delete().then(() => {
+                    // Remove a linha correspondente na tabela
+                    const row = button.closest('tr');
+                    row.remove();
+                    console.log('Produto removido com sucesso.');
+                }).catch((error) => {
+                    console.log('Erro ao remover o produto:', error);
+                });
+            }
+        } else {
+            console.log('Nenhum distribuidor encontrado com o e-mail fornecido.');
+        }
+    }).catch((error) => {
+        console.error('Erro ao buscar distribuidor:', error);
+    });
 }
+
